änderungen an models und Dokumentation hinzugefügt

+++ carSharingNow/carSharingNow.zip_expanded/carSharingNow/src/main/java/com/csn/carSharingNow/controller/CarController.java
@@ -27,7 +27,7 @@ public class CarController {
         return carList;
     }
 
-    public Car getCarByID(int Id) {
+    public Car getCarByID(Long Id) {
         Car car = carRepository.findCarById(Id);
         return car;
     }
