+++ carSharingNow/carSharingNow.zip_expanded/carSharingNow/src/main/java/com/csn/carSharingNow/repositories/CarRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CarRepository extends JpaRepository<Car, Integer > {
-    Car findCarById(int id);
+public interface CarRepository extends JpaRepository<Car, Long > {
+    Car findCarById(Long id);
 
     List<Car> findCarByCarStationEnum(String CarStationEnum);
